requirements.txt
cohere>=4.0.0
numpy>=1.21.0
pandas>=1.3.0
python-dotenv>=0.19.0
annoy>=1.17.0